requirements.txt
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.13